refactor(world): change most values from ints to doubles (#22)
Instead of getting the whole world, individuals now only get objects
within their vision

BREAKING CHANGE:
* Convert all positions and energy levels from `int` to `doubles`.
* Convert all usages of ID from `int` to `long`.

+++ src/model/model.cpp
@@ -1,6 +1,6 @@
 #include "model.h"
 
-Point::Point(int x, int y) : x{x}, y{y} {}
+Point::Point(double x, double y) : x{x}, y{y} {}
 
 bool Point::operator == (const Point & p) {return x == p.x && y == p.y;}
 double Point::operator / (const Point & p) {return sqrt(pow(p.x - x, 2.0) + pow(p.y - y, 2.0));}
