+++ src/model/object/entity/individual/fred/fred.cpp
@@ -1,6 +1,6 @@
 #include "fred.h"
 
-Fred::Fred(int x, int y, int energy) : Individual(x, y, energy) {}
+Fred::Fred(double x, double y, double energy) : Individual(x, y, energy) {}
 
 Fred::Individual::~Individual() = default;
