+++ src/model/object/entity/individual/fred/fred.h
@@ -7,7 +7,7 @@
 
 class Fred : public Individual {
 public:
-    Fred(int x, int y, int energy);
+    Fred(double x, double y, double energy);
 
     Action * act(std::vector<Object *> * visibles);
 };
